+++ next.config.js
@@ -1,7 +1,15 @@
-const withMDX = require('@next/mdx')({
-  extension: /\.mdx?$/,
-});
+const withMdxEnhanced = require('next-mdx-enhanced');
 
-module.exports = withMDX({
+module.exports = withMdxEnhanced({
+  layoutPath: 'layouts',
+  defaultLayout: true,
+  fileExtensions: ['mdx'],
+  remarkPlugins: [],
+  rehypePlugins: [],
+  extendFrontMatter: {
+    process: (mdxContent, frontMatter) => {},
+    phase: 'prebuild|loader|both',
+  },
+})({
   pageExtensions: ['js', 'jsx', 'md', 'mdx', 'ts', 'tsx'],
 });
