+++ package.json
@@ -9,8 +9,8 @@
   },
   "dependencies": {
     "@mdx-js/loader": "^1.6.0",
-    "@next/mdx": "^9.3.6",
     "next": "9.3.5",
+    "next-mdx-enhanced": "^2.5.0",
     "react": "16.13.1",
     "react-dom": "16.13.1"
   },
