+++ pages/blog/[title].tsx
@@ -1,5 +1,6 @@
 import { withRouter, SingletonRouter } from 'next/router';
 import React from 'react';
+import Layout from "../../components/Layout";
 
 interface Props {
   router: SingletonRouter
@@ -7,7 +8,13 @@ interface Props {
 
 class Post extends React.PureComponent<Props> {
   render() {
-    return (<div />);
+    const { router } = this.props;
+    console.log(router)
+    return (
+      <Layout>
+        <div>1</div>
+      </Layout>
+    );
   }
 }
